cloud-fonction-controleur-motopompe/requirements.txt
google-cloud-bigquery
google-cloud-iot